refactor: Fare 패키지 생성
